# src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: Task API
  description: |
    A RESTful API for managing tasks using a contract-first approach.
    This API demonstrates OpenAPI specification, Flyway migrations, and PostgreSQL integration.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8080/api/v1
    description: Local development server

tags:
  - name: Task
    description: Task management operations

paths:
  /tasks:
    get:
      tags:
        - Task
      summary: Get all tasks
      description: Retrieves a list of all tasks in the system
      operationId: getAllTasks
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    post:
      tags:
        - Task
      summary: Create a new task
      description: Creates a new task with the provided information
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  
  /tasks/{id}:
    get:
      tags:
        - Task
      summary: Get task by ID
      description: Retrieves a specific task by its ID
      operationId: getTaskById
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    put:
      tags:
        - Task
      summary: Update task
      description: Updates an existing task with the provided information
      operationId: updateTask
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    delete:
      tags:
        - Task
      summary: Delete task
      description: Deletes a task by its ID
      operationId: deleteTask
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - description
        - status
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the task
          example: 1
        title:
          type: string
          description: Task title
          minLength: 1
          maxLength: 100
          example: "Finish documentation"
        description:
          type: string
          description: Task description
          maxLength: 500
          example: "Complete the API documentation for the project"
        status:
          type: string
          description: Current status of the task (e.g., 'Pending', 'In Progress', 'Completed')
          enum:
            - Pending
            - In Progress
            - Completed
          example: "In Progress"
        dueDate:
          type: string
          format: date-time
          description: The due date for the task
          example: "2026-03-01T12:00:00Z"
    
    TaskRequest:
      type: object
      required:
        - title
        - description
        - status
      properties:
        title:
          type: string
          description: Task title
          minLength: 1
          maxLength: 100
          example: "Finish documentation"
        description:
          type: string
          description: Task description
          maxLength: 500
          example: "Complete the API documentation for the project"
        status:
          type: string
          description: Current status of the task
          enum:
            - Pending
            - In Progress
            - Completed
          example: "In Progress"
        dueDate:
          type: string
          format: date-time
          description: The due date for the task
          example: "2026-03-01T12:00:00Z"

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        status:
          type: integer
          description: HTTP status code
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error message
        path:
          type: string
          description: Request path